@@ -1,13 +1,13 @@
-const app = require('./app'),
-  Rollbar = require('rollbar'),
-  migrationsManager = require('./migrations'),
-  config = require('./config'),
-  logger = require('./app/logger');
+import Rollbar from 'rollbar';
+import app from './app';
+import migrationsCheck from './migrations';
+import config from './config';
+import logger from './app/logger';
 
 const port = config.common.api.port || 8080;
 
-Promise.resolve()
-  .then(() => migrationsManager.check())
+export default Promise.resolve()
+  .then(() => migrationsCheck())
   .then(() => {
     const rollbar = new Rollbar({
       accessToken: config.common.rollbar.accessToken,
@@ -19,5 +19,6 @@ Promise.resolve()
     app.listen(port);
 
     logger.info(`Listening on port: ${port}`);
+    return app;
   })
   .catch(logger.error);
